pyproject.toml
[tool.poetry]
name = "data-cleaning-framework"
version = "0.0.0"
description = ""
authors = ["Chris Hacker <49451910+christopher-hacker@users.noreply.github.com>"]
readme = "README.md"
packages = [
    { include = "data_cleaning_framework" }
]

[tool.poetry.scripts]
dcf = "data_cleaning_framework.cli:cli"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pandas = "^2.1.2"
pandera = {extras = ["io"], version = "^0.17.2"}
pydantic = {extras = ["email"], version = "^2.4.2"}
pyyaml = "^6.0.1"
tqdm = "^4.66.1"
click = "^8.1.7"
questionary = "^2.0.1"
jinja2 = "^3.1.2"
openpyxl = "^3.1.3"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pylint = "^3.0.2"
jupyter = "^1.0.0"
vulture = "^2.11"
pre-commit = "^3.7.1"
pytest = "^8.2.1"
coverage = "^7.5.3"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
